+++ StreamVideoTests/StreamVideoTestCase.swift
@@ -15,6 +15,11 @@ open class StreamVideoTestCase: XCTestCase {
         streamVideo = StreamVideo.mock(httpClient: httpClient)
     }
     
+    override open func tearDown() async throws {
+        try await super.tearDown()
+        await streamVideo?.disconnect()
+    }
+    
     // TODO: replace this with something a bit better
     func waitForCallEvent(nanoseconds: UInt64 = 500_000_000) async throws {
         try await Task.sleep(nanoseconds: nanoseconds)