+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### ✅ Added
 - New API that allows adding/removing capturePhotoOutput, videoOutput on the local participant's AVCaptureSession. [#301](https://github.com/GetStream/stream-video-swift/pull/301)
 - New API that allows zooming the local participant's camera. [#301](https://github.com/GetStream/stream-video-swift/pull/301)
+- VideoFilters for blurring or setting an image as background. [#309](https://github.com/GetStream/stream-video-swift/pull/309)
 
 # [0.5.0](https://github.com/GetStream/stream-video-swift/releases/tag/0.5.0)
 _February 15, 2024_