+++ CHANGELOG.md
@@ -4,6 +4,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 # Upcoming
 
+### 🔄 Changed
+
+# [0.5.1](https://github.com/GetStream/stream-video-swift/releases/tag/0.5.1)
+_March 05, 2024_
+
 ### ✅ Added
 - New API that allows adding/removing capturePhotoOutput, videoOutput on the local participant's AVCaptureSession. [#301](https://github.com/GetStream/stream-video-swift/pull/301)
 - New API that allows zooming the local participant's camera. [#301](https://github.com/GetStream/stream-video-swift/pull/301)